pyproject.toml
[tool.poetry]
name = "db-2024"
version = "0.1.0"
description = ""
authors = ["Piotr Marecki <pmarecki.pro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.9.3"
loguru = "^0.7.2"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
pydantic = "^2.6.4"
argon2-cffi = "^23.1.0"
aiohttp-cors = "^0.7.0"
requests = "^2.31.0"
retry = "^0.9.2"
faker = "^25.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.2"
pytest-aiohttp = "^1.0.5"
pytest-asyncio = "^0.23.6"
requests = "^2.31.0"
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.3"
flake8 = "^7.0.0"
black = "^24.4.2"


[tool.poe.tasks.test]
help = "Run tests"
cmd = "pytest"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
